refactor: move host-agnostic error helpers from common to utils
toErrorMessage/ensureError/extractErrorMessage are used from webview
frontend (browser) code, but lived in src/common, which CLAUDE.md
documents as backend-only. Two frontend files (TexraDiffView.ts,
baseLogFormatter.ts) were reaching into common/errors to get them,
violating that boundary.

Move the implementation to src/utils/errors, the directory actually
shared between the extension host and webview frontends, and leave a
re-export shim at the old location so the ~190 existing backend
consumers of @common/errors and @common/errors/errorMessage are
unaffected.
