+++ src/common/errors/errorMessage.ts
@@ -1,31 +1,13 @@
-import { isNonEmptyString } from '@utils/core';
-
-/** Normalize any thrown value into a user-friendly error message string. */
-export function toErrorMessage(err: unknown): string {
-  if (err instanceof Error) {
-    return err.message;
-  }
-  return String(err);
-}
-
-/** Coerce any thrown value to an Error instance. */
-export function ensureError(err: unknown): Error {
-  return err instanceof Error ? err : new Error(toErrorMessage(err));
-}
-
 /**
- * Extract error message from Error objects or strings.
- * Returns `undefined` if the value is neither an Error nor a non-empty string.
- *
- * Use this when you need optional extraction. For a guaranteed string, use
- * {@link toErrorMessage}.
+ * `toErrorMessage`/`ensureError`/`extractErrorMessage` are host-agnostic and
+ * used from webview frontend (browser) code as well as the extension host,
+ * so they live canonically in `@utils/errors/errorMessage` (the shared
+ * location) rather than here in `common`, which is backend-only. Re-exported
+ * for the existing `@common/errors` and `@common/errors/errorMessage`
+ * consumers.
  */
-export function extractErrorMessage(err: unknown): string | undefined {
-  if (err instanceof Error && isNonEmptyString(err.message)) {
-    return err.message.trim();
-  }
-  if (isNonEmptyString(err)) {
-    return err.trim();
-  }
-  return undefined;
-}
+export {
+  toErrorMessage,
+  ensureError,
+  extractErrorMessage,
+} from '@utils/errors/errorMessage';
